+++ batcave/commander.py
@@ -44,10 +44,12 @@ class SubParser:
 class Commander:
     """This class provides a simplified interface to the argparse.ArgumentParser class."""
 
-    def __init__(self, description: str, arguments: Sequence[Argument] = tuple(), subparsers: Iterable[SubParser] = tuple(),  # pylint: disable=too-many-locals,too-many-arguments
-                 subparser_common_args: Sequence[Argument] = tuple(), default: Optional[Callable] = None, parents: Sequence[ArgumentParser] = tuple(),
-                 parse_extra: bool = False, extra_var_sep: str = ':', convert_extra: bool = True, version: Optional[AppVersion] = None,
-                 version_style: VersionStyle = VersionStyle.one_line, formatter_class: Type[HelpFormatter] = ArgumentDefaultsHelpFormatter):
+    def __init__(self, description: str, arguments: Sequence[Argument] = tuple(),  # pylint: disable=too-many-locals,too-many-arguments,too-many-positional-arguments
+                 subparsers: Iterable[SubParser] = tuple(), subparser_common_args: Sequence[Argument] = tuple(),
+                 default: Optional[Callable] = None, parents: Sequence[ArgumentParser] = tuple(),
+                 parse_extra: bool = False, extra_var_sep: str = ':', convert_extra: bool = True,
+                 version: Optional[AppVersion] = None, version_style: VersionStyle = VersionStyle.one_line,
+                 formatter_class: Type[HelpFormatter] = ArgumentDefaultsHelpFormatter):
         """
         Args:
             description: The subcommand name for the subparser.