+++ batcave/servermgr.py
@@ -15,7 +15,7 @@ from platform import node
 from socket import getfqdn, gethostbyname, gaierror
 from string import Template
 from time import sleep
-from typing import cast, List, Optional, Tuple, TypeAlias, Union
+from typing import cast, List, Optional, Tuple, Union
 from xml.etree.ElementTree import Element, SubElement, parse as xml_parse
 
 # Import third-party modules
@@ -52,9 +52,9 @@ ServiceState = Enum('ServiceState', ('StartPending', 'ContinuePending', 'Running
 ServiceType = Enum('ServiceType', ('systemd', 'sysv', 'upstart', 'windows'))
 TaskSignal = Enum('TaskSignal', ('enable', 'disable', 'run', 'end'))
 
-ServerType: TypeAlias = Union[str, 'Server']
-ServerManager: TypeAlias = Union[WMI, 'OSManager']
-WMIObject: TypeAlias = Union[bool, WMI]
+type ServerType = Union[str, 'Server']
+type ServerManager = Union[WMI, 'OSManager']
+type WMIObject = Union[bool, WMI]
 
 
 class ServerObjectManagementError(BatCaveException):
@@ -918,7 +918,7 @@ class ManagementObject:
             value: The value of the value argument.
         """
         self.type = self.OBJECT_PREFIX + type(self).__name__
-        self.object_ref = object_ref
+        self.object_ref: Optional[ManagementObject] = object_ref
         self.manager = manager
         self.key = key
         self.value = value
@@ -945,7 +945,7 @@ class ManagementObject:
             ServerObjectManagementError.NOT_UNIQUE: If the object is not unique.
         """
         if self.object_ref:
-            if len(results := getattr(self.manager, self.type)(**{self.key: self.value}, **self.key_values)) > 1:
+            if len(results := cast(List[ManagementObject], getattr(self.manager, self.type)(**{self.key: self.value}, **self.key_values))) > 1:
                 raise ServerObjectManagementError(ServerObjectManagementError.NOT_UNIQUE, type=self.type, key=self.key, val=self.value)
             self.object_ref = results[0] if results else None